Rename/expose class
Planning to use it within a machine-level view
